backend/requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
python-dotenv==1.0.1
SQLAlchemy==2.0.38
psycopg2-binary==2.9.10
passlib==1.7.4
python-multipart==0.0.20
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
email_validator==2.2.0
PyJWT==2.10.1
python-telegram-bot==22.0